Enemy Takes Damage, Difficulties
Enemy have a health value which is set depending on the difficulty. An enum was created for the different difficulties in the GameData and a switch was implemented to set different health in different difficulties. Player can shoot the enemies and kill them. Enemies detection range is increased when player is running. Arranged minor things in the scene. Enemy is not able to kill player at the this point.