Changed Player Controller, Added Main Menu/Win/Lose states for Player
Changed the old player controller as it was not compatible with levels which require going up or down slopes. The new game controller is not the best but will manage given that it is a free asset. Menus were also create; Main Menu, Game Over and Victory Menu. The player can navigate through them.